[Fix #155856578] Fix login component
