+++ package.json
@@ -12,7 +12,7 @@
     "eslint": "^4.18.2",
     "expect": "^23.0.0-alpha.0",
     "jest-enzyme": "^6.0.0",
-    "material-ui": "^1.0.0-beta.42",
+    "material-ui": "^0.20.0",
     "mocha": "^5.0.4",
     "react": "^16.2.0",
     "react-bootstrap": "^0.32.1",
