+++ public/index.html
@@ -12,7 +12,8 @@
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
 
     <!-- material-ui fonts -->
-    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+   
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500">
     
     <!-- Required dependencies -->
     <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/prop-types/15.6.0/prop-types.min.js"></script>
