+++ src/pages/LoginComponent.js
@@ -4,6 +4,7 @@ import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import AppBar from 'material-ui/AppBar';
 import RaisedButton from 'material-ui/RaisedButton';
 import TextField from 'material-ui/TextField';
+import LoginForm from '../components/LoginFormComponent';
 
 
 class Login extends Component {
@@ -16,7 +17,8 @@ class Login extends Component {
 
     }
     
-    handleClick(event) {
+    handleClick = event => {
+        console.log("username : qwertgh", this.state)
         var apiBaseUrl = "https://brighter-event.herokuapp.com/api/v1/auth/";
         
         const requestPromise = require('request-promise');
@@ -39,29 +41,16 @@ class Login extends Component {
         
     }
 
+    onChange = event => {
+         this.setState({ [event.target.name]: event.target.value });
+        console.log("current state ", this.state);
+    }
+
     render() {
         return (
             <div>
                 <MuiThemeProvider>
-                    <div>
-                        <AppBar
-                            title="Login"
-                        />
-                        <TextField
-                            hintText="Enter your Username"
-                            floatingLabelText="Username"
-                            onChange={(event, newValue) => this.setState({ username: newValue })}
-                        />
-                        <br />
-                        <TextField
-                            type="password"
-                            hintText="Enter your Password"
-                            floatingLabelText="Password"
-                            onChange={(event, newValue) => this.setState({ password: newValue })}
-                        />
-                        <br />
-                        <RaisedButton label="Submit" primary={true} style={style} onClick={(event) => this.handleClick(event)} />
-                    </div>
+                    <LoginForm handleClick = {this.handleClick} onChange = {this.onChange}/>
                 </MuiThemeProvider>
             </div>
         );
